Tests author routes and functions

+++ back_end/src/controllers/author.controller.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
 import { PrismaClient } from '@prisma/client';
+import { AuthenticatedRequest } from '../middlewares/protect'
 
 
 const prisma = new PrismaClient();
@@ -24,28 +25,48 @@ export const findOne: RequestHandler = async (req, res) => {
   res.json(author);
 };
 
-/**
- * POST /authors
+/**GET /authors/search
+ *
  */
-export const create: RequestHandler = async (req, res) => {
-  const { firstName, lastName, pseudo, email, password, bio, link } = req.body;
-  const newAuthor = await prisma.author.create({
-    data: { firstName, lastName, pseudo, email, password, bio, link }
-  });
-  res.status(201).json(newAuthor);
-};
+export const findByPublicInfo: RequestHandler = async (req, res) => {
+  const { pseudo, nom, prenom } = req.query;
 
+  const authors = await prisma.author.findMany({
+    where: {
+      AND: [
+        pseudo ? { pseudo: { contains: String(pseudo), mode: 'insensitive' } } : {},
+        nom ? { lastName: { contains: String(nom), mode: 'insensitive' } } : {},
+        prenom ? { firstName: { contains: String(prenom), mode: 'insensitive' } } : {},
+      ],
+    },
+  });
+
+  if (!authors.length) {
+    res.status(404).json({ error: 'Aucun auteur trouvé' });
+    return;
+  }
+
+  res.json(authors);
+};
 /**
  * PUT /authors/:id
  */
 export const update: RequestHandler = async (req, res) => {
   const { id } = req.params;
-  const { firstName, lastName, pseudo, email, bio, link } = req.body;
-  const updated = await prisma.author.update({
-    where: { id },
-    data: { firstName, lastName, pseudo, email, bio, link }
-  });
-  res.json(updated);
+	const { sub, roles } = (req as AuthenticatedRequest).user;
+	const isAdmin = roles.includes('admin');
+
+	if (!isAdmin && id !== sub) {
+		res.status(403).json({ error: 'Accès refusé : vous devez agir sur votre propre profil'});
+		return;
+	}
+
+	const { firstName, lastName, pseudo, email, bio, link } = req.body;
+	const updated = await prisma.author.update({
+		where: { id },
+		data: { firstName, lastName, pseudo, email, bio, link }
+	});
+	res.json(updated);
 };
 
 /**
@@ -53,6 +74,14 @@ export const update: RequestHandler = async (req, res) => {
  */
 export const remove: RequestHandler = async (req, res) => {
   const { id } = req.params;
-  await prisma.author.delete({ where: { id } });
-  res.sendStatus(204);
+	const { sub, roles } = (req as AuthenticatedRequest).user;
+	const isAdmin = roles.includes('admin');
+
+	if (!isAdmin && id !== sub) {
+		res.status(403).json({ error: 'Vous pouvez uniquement supprimer votre profil !'});
+		return;
+	}
+
+	await prisma.author.delete({ where: { id } });
+	res.sendStatus(204);
 };
